add FFTW_LIB link to US_PP in on, negf and tddft. Now they can compiled on BW. Before they can compile on my desktop but has linking problems on BW
git-svn-id: svn+ssh://chips.ncsu.edu/home/svn/repos/codes@3090 eb38b3b0-33ea-0310-94ad-d43fea0d214f

+++ Headers/rmg_alloc.h
@@ -16,13 +16,13 @@ void my_free( void *ptr );
 
 #define my_free(_ptr_) {rmg_free(_ptr_); _ptr_ = NULL;}
 #define my_malloc( _ptr_, _nobj_, _type_ ) \
-        _ptr_ = rmg_malloc((int) _nobj_, sizeof(_type_))
+        _ptr_ = rmg_malloc((size_t) _nobj_, sizeof(_type_))
 
 #define my_calloc( _ptr_, _nobj_, _type_ ) \
-        _ptr_ = rmg_calloc((int) _nobj_, sizeof(_type_))
+        _ptr_ = rmg_calloc((size_t) _nobj_, sizeof(_type_))
 
 #define my_malloc_init( _ptr_, _nobj_, _type_ ) \
-        _ptr_ = rmg_malloc_init((int) _nobj_, sizeof(_type_), #_type_)
+        _ptr_ = rmg_malloc_init((size_t) _nobj_, sizeof(_type_), #_type_)
 
 
 #ifdef __cplusplus
